Updated acceptance tests to separate tpm simulator from hirs provisioning
