+++ spec/acceptance/nodesets/default.yml
@@ -31,18 +31,18 @@ HOSTS:
 #        mirrorlist: 'https://mirrors.fedoraproject.org/metalink?repo=epel-7&arch=$basearch'
 #        gpgkeys:
 #          - https://getfedora.org/static/352C64E5.txt
-#  el6-tpm-1-2:
-#    roles:
-#      - hirs
-#      - tpm_1_2
-#    platform:   el-6-x86_64
-#    box:        centos/6
-#    hypervisor: <%= hypervisor %>
-#    yum_repos:
-#      epel:
-#        mirrorlist: 'https://mirrors.fedoraproject.org/metalink?repo=epel-6&arch=$basearch'
-#        gpgkeys:
-#          - https://getfedora.org/static/0608B895.txt
+  el6-tpm-1-2:
+    roles:
+      - hirs
+      - tpm_1_2
+    platform:   el-6-x86_64
+    box:        centos/6
+    hypervisor: <%= hypervisor %>
+    yum_repos:
+      epel:
+        mirrorlist: 'https://mirrors.fedoraproject.org/metalink?repo=epel-6&arch=$basearch'
+        gpgkeys:
+          - https://getfedora.org/static/0608B895.txt
   aca:
     roles:
       - aca
